feat: add car status object generation feature
adds a feature that generates a car status object.
Each key of the object is a car's name,
and the value is the number of times that car has moved forward.

These forward counts are calculated
by the generateRandomNumbersByAttempts method, which generates
a specific number of random numbers
and counts how many are 4 or greater.